+++ pkg/secrets/filestore.go
@@ -14,17 +14,5 @@
 
 package secrets
 
-import (
-	v1 "k8s.io/api/core/v1"
-)
-
 // FilestoreSecretType represents the secret type for Filestore credentials.
 const FilestoreSecretType string = "secrets.kanister.io/filestore"
-
-// ValidateFileStoreCredentials validates secret has all necessary information
-// for Filestore Credentials
-func ValidateFileStoreCredentials(secret *v1.Secret) error {
-	// Currently we dont need credentials for filestore hence
-	// keeping the validation empty
-	return nil
-}
