+++ pkg/secrets/repositoryserver/const.go
@@ -47,18 +47,4 @@ const (
 	RepoPasswordKey  = "repo-password"
 	AdminUsernameKey = "username"
 	AdminPasswordKey = "password"
-
-	// Error msg for missing required field in the secret
-	MissingRequiredFieldErrorMsg = "Missing required field %s in the secret '%s:%s'"
-	// Error msg for unknown in the secret
-	UnknownFieldErrorMsg = "'%s:%s' secret has an unknown field"
-	// Unsupported location type in the secret
-	UnsupportedLocationTypeErrorMsg = "Unsupported location type '%s' for secret '%s:%s'"
-	// Invalid Secret type error msg
-	IncompatibleSecretTypeErrorMsg = "Incompatible secret type. Expected type %s in the secret '%s:%s'"
-
-	// Nil Secret error message
-	NilSecretErrorMessage = "Secret is Nil"
-	// Empty Secret error message
-	EmptySecretErrorMessage = "Empty secret. Expected at least one key value pair in the secret '%s:%s'"
 )
