+++ pkg/secrets/repositoryserver/repository_server_admin_credentials.go
@@ -17,6 +17,8 @@ package repositoryserver
 import (
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
+
+	secerrors "github.com/kanisterio/kanister/pkg/secrets/errors"
 )
 
 var _ Secret = &repositoryServerAdminCredentials{}
@@ -33,21 +35,21 @@ func NewRepositoryServerAdminCredentials(secret *corev1.Secret) *repositoryServe
 
 func (r *repositoryServerAdminCredentials) Validate() error {
 	if r.credentials == nil {
-		return errors.Wrapf(ErrValidate, NilSecretErrorMessage)
+		return errors.Wrapf(secerrors.ErrValidate, secerrors.NilSecretErrorMessage)
 	}
 	if len(r.credentials.Data) == 0 {
-		return errors.Wrapf(ErrValidate, EmptySecretErrorMessage, r.credentials.Namespace, r.credentials.Name)
+		return errors.Wrapf(secerrors.ErrValidate, secerrors.EmptySecretErrorMessage, r.credentials.Namespace, r.credentials.Name)
 	}
 
 	// kopia repository server admin credentials secret must have exactly 2 fields
 	if len(r.credentials.Data) != 2 {
-		return errors.Wrapf(ErrValidate, UnknownFieldErrorMsg, r.credentials.Namespace, r.credentials.Name)
+		return errors.Wrapf(secerrors.ErrValidate, secerrors.UnknownFieldErrorMsg, r.credentials.Namespace, r.credentials.Name)
 	}
 	if _, ok := r.credentials.Data[AdminUsernameKey]; !ok {
-		return errors.Wrapf(ErrValidate, MissingRequiredFieldErrorMsg, AdminUsernameKey, r.credentials.Namespace, r.credentials.Name)
+		return errors.Wrapf(secerrors.ErrValidate, secerrors.MissingRequiredFieldErrorMsg, AdminUsernameKey, r.credentials.Namespace, r.credentials.Name)
 	}
 	if _, ok := r.credentials.Data[AdminPasswordKey]; !ok {
-		return errors.Wrapf(ErrValidate, MissingRequiredFieldErrorMsg, AdminPasswordKey, r.credentials.Namespace, r.credentials.Name)
+		return errors.Wrapf(secerrors.ErrValidate, secerrors.MissingRequiredFieldErrorMsg, AdminPasswordKey, r.credentials.Namespace, r.credentials.Name)
 	}
 	return nil
 }
